# docs/swagger.yaml
definitions:
  manufacturer.AnomalyDTO:
    properties:
      anomalyType:
        default: stuck
        type: string
    type: object
  manufacturer.CertificateDTO:
    properties:
      certificateId:
        type: string
    type: object
  manufacturer.ConnectionDTO:
    properties:
      connectionModel:
        additionalProperties: {}
        type: object
      connectionType:
        default: simple-CoAP
        type: string
    type: object
  manufacturer.ControllCommandDTO:
    properties:
      serviceDefinition:
        default: lamp
        type: string
      serviceUri:
        default: /lamp
        type: string
    type: object
  manufacturer.DigitalTwinDTO:
    properties:
      certificateId:
        type: string
      controlCommands:
        items:
          $ref: '#/definitions/manufacturer.ControllCommandDTO'
        type: array
      handleableAnomalies:
        items:
          $ref: '#/definitions/manufacturer.AnomalyDTO'
        type: array
      physicalTwinConnection:
        $ref: '#/definitions/manufacturer.ConnectionDTO'
      sensedProperties:
        items:
          $ref: '#/definitions/manufacturer.SensedPropertyDTO'
        type: array
      systemName:
        type: string
    type: object
  manufacturer.SensedPropertyDTO:
    properties:
      intervalTime:
        default: 10
        type: integer
      sensorEndpointMode:
        default: INTERVAL_RETRIEVAL
        type: string
      serviceDefinition:
        default: temperature
        type: string
      serviceUri:
        default: /temperature
        type: string
    type: object
  manufacturer.SystemDefinitionDTO:
    properties:
      address:
        default: localhost
        type: string
      authenticationInfo:
        type: string
      port:
        default: 5000
        type: integer
      systemName:
        default: my-digital-twin
        type: string
    type: object
info:
  contact:
    url: https://github.com/MrDweller/digital-twin-hub
  description: This page shows the REST interfaces offered by the Digital Twin Hub.
  title: Digital Twin Hub
  version: "1.0"
paths:
  /create-digital-twin:
    post:
      description: Create a new digital twin based on the given JSON object. This
        will create a connection to the physical twin based on the connection info
        given, this will also include generating endpoints to controll and view sensed
        data.
      parameters:
      - description: DigitalTwinDTO JSON
        in: body
        name: DigitalTwin
        required: true
        schema:
          $ref: '#/definitions/manufacturer.DigitalTwinDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/manufacturer.SystemDefinitionDTO'
      summary: Create a new digital twin
      tags:
      - Management
  /remove-digital-twin:
    delete:
      description: Delete a digital twin based on the given address and port.
      parameters:
      - description: address
        in: query
        name: address
        required: true
        type: string
      - description: 'port '
        in: query
        name: port
        required: true
        type: string
      responses:
        "200":
          description: OK
      summary: Delete a digital twin
      tags:
      - Management
  /upload-certificates:
    post:
      description: Upload certificate files as zip, to be used by a digital twin.
        Takes cert.pem and key.pem files and gives `certId`.
      parameters:
      - description: Cert file
        in: formData
        name: cert
        required: true
        type: file
      - description: Key file
        in: formData
        name: key
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/manufacturer.CertificateDTO'
      summary: Upload certificate files as zip
      tags:
      - Management
swagger: "2.0"